Adjust for NPM install differences in common.js
Previously, if the samples were run when installed from an SDK
package on NPM v2 (default NPM version for Node.js 4.x), the
samples could fail on a `require` for an SDK dependency which
could not be found. NPM v2 nests installed dependencies,
preventing any code running outside of the SDK lib directory
from being able to find the nested dependency properly.
`require` calls for nested dependencies installed via NPM v3 and
later would not fail since the nested dependencies are flattened
into the root node_modules directory.

This commit changes the `requireClient` function in `common.js`
to be named `require`. The implementation of `require` now
handles the ability to require the SDK package by explicit name
when running either from source or a package install. The
implementation also adjusts the module path if needed in order
to resolve a dependency installed for various NPM versions.